Add aws_s3 and WazuhIntegration files
